"some simplifications and optimizations of multiblade; updates #123"

+++ prototype2/multiblade/common/EventBuilder.h
@@ -36,10 +36,10 @@
 
 namespace Multiblade {
 
-/// Struct containing channel number and ADC-value for a singe strip or wire.
+/// Struct containing channel number and ADC-value for a single strip or wire.
 struct point {
-  uint32_t channel; ///< Number of either wire or strip channel
-  uint32_t ADC;     ///< Value of the signal from the corresponding channel
+  uint16_t channel; ///< Number of either wire or strip channel
+  uint16_t ADC;     ///< Value of the signal from the corresponding channel
 };
 
 /// Overloaded < operator. Required when we need to sort the data points for
@@ -75,7 +75,7 @@ public:
   /// \param ADC Signal value
   /// \param clock Clock-cycle number
   /// \return True when the cluster/event is complete, false otherwise
-  bool addDataPoint(const uint8_t &channel, const uint16_t &ADC,
+  bool addDataPoint(const uint16_t &channel, const uint16_t &ADC,
                     const uint32_t &clock);
 
   /// Call this function when last point of the run has been received and the
@@ -144,7 +144,7 @@ public:
   /// channels will be considered invalid.
   ///
   /// \param nchannels
-  void setNumberOfWireChannels(uint8_t nchannels) {
+  void setNumberOfWireChannels(uint16_t nchannels) {
     m_nwire_channels = nchannels;
   }
 
@@ -153,7 +153,7 @@ public:
   /// channels will be considered invalid.
   ///
   /// param nchannels
-  void setNumberOfStripChannels(uint8_t nchannels) {
+  void setNumberOfStripChannels(uint16_t nchannels) {
     m_nstrip_channels = nchannels;
   }
 
@@ -199,9 +199,9 @@ private:
   /// Time window for a single cluster. In number of clock-cycles
   uint32_t m_time_window{400};
   /// Number of wire channels
-  uint8_t m_nwire_channels{32};
+  uint16_t m_nwire_channels{32};
   /// Number of strip channels
-  uint8_t m_nstrip_channels{32};
+  uint16_t m_nstrip_channels{32};
   /// Whether to use wieghted average or not when calculation position
   bool m_use_weighted_average{true};
 
@@ -243,7 +243,7 @@ private:
   ///
   /// \param channel Channel number
   /// \param ADC Signal value
-  void addPointToCluster(uint32_t channel, uint32_t ADC);
+  void addPointToCluster(uint16_t channel, uint16_t ADC);
 
   /// This function is called once a cluster is formed
   bool processClusters();
